ci: separate deployable check from next deployment script
