+++ .github/workflows/deploy-next.yml
@@ -31,6 +31,9 @@ jobs:
           GH_TOKEN_FOR_STORYBOOK: ${{ github.actor }}:${{ secrets.ADMIN_TOKEN }}
         run: |
           if [ "$NEXT_RELEASE_ENABLED" == "true" ]; then
+            npm run util:is-next-deployable
+            # deploy storybook, but still release next if it fails
+            { npm run build-storybook && npx storybook-to-ghpages --host-token-env-variable=GH_TOKEN_FOR_STORYBOOK --existing-output-dir=docs --ci; } || true
             npm run util:deploy-next-from-ci
           else
             echo "Next release is disabled"
